Move log10 from core.helpers to helpers.math
